PB-676: simplify yq filter.
The comment about "-r" not printing only the value is wrong. The problem was
that the yq expression was returning a list. With this new expression we get
the value within the list and it shows only what we want. No more need for awk.
Also the regex is simplified a little.