PB-676: pin yq version to 4.44.1.
Before this change we use the latest upstream version of yq. We have had issues
with upstream packages disappearing under us (see PB-300). Also yq has
historically had backwards incompatible changes. This change ensures we use a
well known version of yq and it remains easy to update it in a controlled way.