PB-992: add support for partitioned tables
The list of indexes to be created and dropped is built with pg_dump and
the name of the table as -t "table-name" parameter. By default this is
not dumping the Index Creation/Drop statements for paritioned table
children. By changing the pg_dump paramater to -t "table-name*" index
creation/drop statements of the paritioned children tables are also
printed out.

additionally a check has been added. table deploy is only possible if
the partitions of the source and the target table are the same